requirements.txt
numpy==1.18.1
pandas==0.25.3
scipy==1.4.1
scikit-learn==0.22.1
seaborn==0.10.1
matplotlib==3.1.2
tqdm==4.41.1
joblib==0.14.1
networkx==2.4
statsmodels==0.10.2
cvxpy==1.0.25
grakel-dev==0.1a6
networkx==2.4
names==0.3.0
statsmodels==0.10.2
gif==1.0.4
natsort==7.0.1
xlrd==1.2.0
lifelines==0.24.16
